% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixeltrix.R
\name{draw_pixels}
\alias{draw_pixels}
\title{BACKEND - draw_pixels()}
\usage{
draw_pixels(m, colours = c("white", "forestgreen"))
}
\arguments{
\item{m}{The matrix that you want to visualize.}

\item{colours}{An optional vector of two colours for 0s and 1s in the
matrix.}
}
\value{
The plot is drawn - so I guess this is what is "returned" by this
 function.
}
\description{
This is a backend function.
DO NOT USE IT.
The only function in the livelycells package for a use is play()!
}
\details{
This hidden function visualizes a matrix as a plot.
}
