% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{evolve}
\alias{evolve}
\title{BACKEND - evolve()}
\usage{
evolve(input_matrix, rule_string)
}
\arguments{
\item{input_matrix}{Matrix in question.}

\item{rule_string}{in Golly/RLE format (e.g., "B3/S23")}
}
\value{
Returns a list of 2 elements. First is the matrix after
 evolution. Second is whether the border has been reached.
 If border_reached then input_matrix is simply returned.
}
\description{
This is a backend function.
DO NOT USE IT.
The only function in the livelycells package for a use is play()!
}
\details{
Evolves a matrix to its next generation according to a rule set.
}
