% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{extract_rules}
\alias{extract_rules}
\title{BACKEND - extract_rules()}
\usage{
extract_rules(input_string)
}
\arguments{
\item{input_string}{The string in Folly/RLE format}
}
\value{
Returns a list of two elements. First element contains birth_rules.
 Second contains survival_rules.
}
\description{
This is a backend function.
DO NOT USE IT.
The only function in the livelycells package for a use is play()!
}
\details{
Takes a string of rules in Golly/RLE format as input
and returns a list with two vectors.
}
