% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livelycells.R
\docType{package}
\name{livelycells}
\alias{livelycells}
\title{An R Shiny App for Conway´s Game of Life}
\description{
The package has one function.
}
\section{Play}{

\emph{play} starts the app so you can then play around with the lively cells.
}

