% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{neighbours}
\alias{neighbours}
\title{BACKEND - neighbours()}
\usage{
neighbours(matrix, row_index, col_index)
}
\arguments{
\item{matrix}{Matrix in question.}

\item{row_index}{of the cell.}

\item{colum_index}{of the cell.}
}
\value{
Returns a value between 0 and 8.
}
\description{
This is a backend function.
DO NOT USE IT.
The only function in the livelycells package for a use is play()!
}
\details{
Counts living neighbours in a cell´s Moore neighbourhood (the surrounding 8)
}
