% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Starts the App to Play With}
\usage{
play()
}
\arguments{
\item{none}{There are no arguments!}
}
\value{
The app just starts in your browser. So, nothing is returned, I guess?
}
\description{
Run this function to start the app in your browser and play the Game.
}
\details{
For more details on the Game see its \href{https://en.wikipedia.org/wiki/Conway's_Game_of_Life}{Wikipedia article}.
}
\examples{
play()

}
